backoffice/requirements.txt
autopep8==1.5.4
bcrypt==3.2.0
cffi==1.14.4
click==7.1.2
cryptography==3.3.1
dnspython==2.0.0
dominate==2.6.0
email-validator==1.1.2
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Bootstrap==3.3.7.1
Flask-Login==0.5.0
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
idna==3.1
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
Pillow==8.1.0
psycopg2-binary==2.8.6
pycodestyle==2.6.0
pycparser==2.20
qrcode==6.1
rope==0.18.0
six==1.15.0
SQLAlchemy==1.3.22
toml==0.10.2
visitor==0.1.3
Werkzeug==1.0.1
WTForms==2.3.3